+++ Cargo.toml
@@ -29,6 +29,7 @@ members = [
     "crates/perl-lsp-transport",
     "crates/perl-lsp-on-type-formatting",
     "crates/perl-lsp-tooling",
+    "crates/perl-lsp-perltidy",
     "crates/perl-lsp-symbol-query",
     "crates/perl-lsp-performance",
     "crates/perl-lsp-critic-parser",
@@ -196,6 +197,7 @@ allow = [
     "perl-lsp-document-links",
     "perl-lsp-workspace-symbols",
     "perl-lsp-tooling",
+    "perl-lsp-perltidy",
     "perl-lsp-performance",
     "perl-lsp-critic-parser",
     "perl-lsp-formatting-types",
@@ -354,6 +356,7 @@ perl-lsp-capability-map = { path = "crates/perl-lsp-capability-map", version = "
 perl-parser-core = { path = "crates/perl-parser-core", version = "0.11.0" }
 perl-lsp-transport = { path = "crates/perl-lsp-transport", version = "0.11.0" }
 perl-lsp-tooling = { path = "crates/perl-lsp-tooling", version = "0.11.0" }
+perl-lsp-perltidy = { path = "crates/perl-lsp-perltidy", version = "0.11.0" }
 perl-lsp-on-type-formatting = { path = "crates/perl-lsp-on-type-formatting", version = "0.11.0" }
 perl-lsp-formatting-types = { path = "crates/perl-lsp-formatting-types", version = "0.11.0" }
 perl-lsp-formatting = { path = "crates/perl-lsp-formatting", version = "0.11.0" }
